HUMANITYSTODOLIST Initial Commit
This is a project that began as mission work attempting to identify immediately actionable tools and methods to help ethnic minorities in the mountains of Burma industrialize and become self sufficient. It has since evolved into a basic reportoire of "how to recreate civilization from scratch", with the premise that all humans are equally valuable and capable of contributing to, modifying, and refining the information here- hence its applicability for an open source project.